Refactor tests and fix get-task
