+++ Justfile
@@ -1,2 +1,2 @@
 test:
-	ruby test/test_start_task.rb
+        ruby -Itest -e 'Dir["test/test_*.rb"].each { |f| require File.expand_path(f) }'
